Convert price to double or approriate according to the given ratio
This is to avoid compilation error with the next commit

+++ examples/mock_exchange_connector.h
@@ -93,7 +93,7 @@ struct MockExchangeConnector : public HttpExchangeConnector {
             auto & resp = current->winningResponse(spotNum);
             result += ML::format("{\"id\":\"%s\",\"max_price\":%lld,\"account\":\"%s\"}",
                              ML::jsonEscape(auction.request->imp.at(spotNum).id.toString()).c_str(),
-                             (long long)(MicroUSD_CPM(resp.price.maxPrice)),
+                             (int64_t)(MicroUSD_CPM(resp.price.maxPrice)),
                              resp.account.toString('.'));
         }
